+++ hoplite-aws/src/test/kotlin/com/sksamuel/hoplite/aws/RegionDecoderTest.kt
@@ -17,13 +17,13 @@ class RegionDecoderTest : StringSpec() {
   init {
     "region converter" {
       RegionDecoder().safeDecode(
-        StringNode("us-east-1", Pos.NoPos, DotPath.root),
+        StringNode("us-east-1", Pos.NoPos, DotPath.root, emptyMap()),
         Region::class.createType(),
         DecoderContext.zero
       ) shouldBe Region.getRegion(Regions.US_EAST_1).valid()
 
       RegionDecoder().safeDecode(
-        StringNode("us-qwewqe-1", Pos.NoPos, DotPath.root),
+        StringNode("us-qwewqe-1", Pos.NoPos, DotPath.root, emptyMap()),
         Region::class.createType(),
         DecoderContext.zero
       ) shouldBe