+++ hoplite-core/src/main/kotlin/com/sksamuel/hoplite/preprocessor/EnvOrSystemPropertyPreprocessor.kt
@@ -2,11 +2,13 @@
 
 package com.sksamuel.hoplite.preprocessor
 
+import com.sksamuel.hoplite.CommonMetadata
 import com.sksamuel.hoplite.ConfigResult
 import com.sksamuel.hoplite.Node
 import com.sksamuel.hoplite.PrimitiveNode
 import com.sksamuel.hoplite.StringNode
 import com.sksamuel.hoplite.fp.valid
+import com.sksamuel.hoplite.withMeta
 
 /**
  * Replaces strings of the form ${var} with the value of the env variable or system property 'var'.
@@ -22,7 +24,8 @@ object EnvOrSystemPropertyPreprocessor : TraversingPrimitivePreprocessor() {
 
   override fun handle(node: PrimitiveNode): ConfigResult<Node> = when (node) {
     is StringNode -> {
-      val value = regex.replace(node.value) { match ->
+      val rawValue = node.value
+      val value = regex.replace(rawValue) { match ->
         val key = match.groupValues[1]
         when (val matchWithDefault = valueWithDefaultRegex.matchEntire(key)) {
           null -> System.getProperty(key) ?: System.getenv(key) ?: match.value
@@ -34,7 +37,7 @@ object EnvOrSystemPropertyPreprocessor : TraversingPrimitivePreprocessor() {
           }
         }
       }
-      node.copy(value = value).valid()
+      node.copy(value = value).withMeta(CommonMetadata.UnprocessedValue, rawValue).valid()
     }
     else -> node.valid()
   }