+++ changelog.md
@@ -5,6 +5,13 @@
 * Added support for AWS Secrets Manager using Amazon's SDK version 2. The syntax is the same, but import `hoplite-aws2` instead of `hoplite-aws`.
 * Added CascadeMode to control how config should cascade through files.
 
+**Breaking Changes**
+
+These breaking changes are to advanced customization features. Most users will need to change no code.
+
+* `ReporterBuilder` has been removed - specify the _obfuscator_ and _secrets policy_ directly on the `ConfigLoaderBuilder`.
+* `SecretsPolicy` interface has been changed to directly use `Node`s and not just `Path`s.
+
 ### 2.5.2
 
 * Fixed sealed class ordering in Kotlin 1.7.x #331